requirements-dev.txt
certifi==2020.4.5.1
dj-static==0.0.6
Django==2.2.4
djangorestframework==3.11.0
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
static3==0.7.0
